chore(code-builder): [collision] remove unused source files
