+++ modules/code-builder/src/collision/index.ts
@@ -1,2 +1,2 @@
 export { default as CollisionSpaceBrute } from './Brute';
-export { default as CollisionSpaceQuadTree } from './QTree';
+export { default as CollisionSpaceQuadTree } from './QuadTree';
